+++ src/intranet3/__init__.py
@@ -118,6 +118,9 @@ def main(global_config, **settings):
     pyramid_config.add_route('api_presence', 'api/presence')
     pyramid_config.add_route('api_blacklist', 'api/blacklist')
     pyramid_config.add_route('deferred', '/_i/deferred')
+    pyramid_config.add_route('api_lateness', '/api/lateness')
+    pyramid_config.add_route('api_absence', '/api/absence')
+    pyramid_config.add_route('api_absence_days', '/api/absence_days')
 
     pyramid_config.add_renderer('.html', 'pyramid_jinja2.renderer_factory')
     pyramid_config.add_renderer(None, 'intranet3.utils.renderer.renderer_factory')