+++ src/intranet3/forms/employees.py
@@ -60,15 +60,18 @@ class LateApplicationForm(BaseForm):
         self.user = kwargs.pop('user')
         super(LateApplicationForm, self).__init__(*args, **kwargs)
 
+    def validate_late_start(self, field):
+        if not field.data:
+            raise ValidationError(_(u'Cannot accept midnight'))
+        if field.data >= self.late_end.data:
+            raise ValidationError(_(u'Invalid time range'))
+
+    def validate_late_end(self, field):
+        if not field.data:
+            raise ValidationError(_(u'Cannot accept midnight'))
+
     def validate_popup_date(self, field):
         date = field.data
-        user_id = self.user.id
-        late = Late.query.filter(Late.date==date)\
-                         .filter(Late.user_id==user_id)\
-                         .filter(Late.deleted==False).first()
-        if late:
-            raise ValidationError(_(u'You already have application in this date'))
-
         if date < datetime.date.today():
             start_date = datetime.datetime.combine(date, day_start)
             end_date = datetime.datetime.combine(date, day_end)