+++ src/intranet3/utils/views.py
@@ -127,10 +127,16 @@ class ApiView(BaseView):
 
             return response
         request.add_response_callback(_response_exception)
-        
+
         super(ApiView, self).__init__(context, request)
 
         self.flash = lambda message, klass='': None # We don't need flash messages. So do nothing
+
+    def _note_presence(self):
+        """
+        No presence checking for calls to API
+        """
+        pass
 
     def dispatch(self):
         if self.request.method.lower() in self.http_method_names: