+++ CHANGES.md
@@ -1,3 +1,20 @@
+0.9.6 (28-11-2013)
+===========
+- better mobile support
+- project configuration allows to add custom tabs in sprint with embedded iframe
+- user date of birth
+
+0.9.5 (17-11-2013)
+===============
+- much better scrum board layout 
+- extra links to /times/tickets/report
+- better out of office form
+- angular sortby polish chart support
+- no team option for user list
+- note presence only for views, not api 
+- teams page visible for users
+- bug fixes
+
 0.9.4 (14-11-2013)
 ===============
 - absence / lates enhancemets